# api/openapi.yaml
# nbpdns's API (ADR-0012, ADR-0033): the source of truth for the API.
# `make generate` generates the server's code and the embedded copy of this
# file from it, and the reference page docs/reference/api.md. Every change
# must stay backward compatible: there's no version in the paths.
openapi: 3.1.0
info:
  title: nbpdns API
  version: 1.0.0
  description: >-
    The read-only API of nbpdns, which compares the DNS data in NetBox's DNS
    plugin with what each PowerDNS server group's primary serves. It serves
    the service's status, each server group's drift, its zones and their
    changes, and each zone's DNS records as NetBox defines them. Everything
    is last-known state, as of the last successful read of each side. The
    API has no authentication until M10, so serve it only on a trusted
    network.
  contact:
    name: nbpdns
    url: https://github.com/zeddD1abl0/netbox-powerdns-ai
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  x-api-id: f4c4a283-c96e-410e-a606-bdda49d9af95
  x-audience: company-internal
servers:
  - url: /api
    description: The API, at the service's address, `server.listen`.
# No authentication until M10 (ADR-0033): an empty requirement says so.
security:
  - {}
tags:
  - name: service
    description: The nbpdns service itself.
  - name: server-groups
    description: >-
      PowerDNS server groups: each is a primary, the servers that follow it,
      and the NetBox views whose zones they serve. A group's name is its
      stable ID.
paths:
  /status:
    get:
      operationId: getStatus
      summary: Get the service's status
      description: >-
        Returns the service's build, uptime, readiness and liveness, the
        refresh schedule and how the last refresh went, NetBox's state, and
        where spans are exported. Each server group's state is a resource
        of its own, under `/server-groups`.
      tags: [service]
      parameters:
        - $ref: "#/components/parameters/FlowID"
      responses:
        "200":
          description: The service's status.
          headers:
            X-Flow-ID:
              $ref: "#/components/headers/FlowID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
              example:
                version: v0.2.0
                revision: ec048d836f57f7ee1102e27b7baa38d8a17469bc
                started: "2026-10-08T01:00:00Z"
                uptime_seconds: 754.2
                ready: true
                live: true
                schedule:
                  interval_seconds: 300
                  timeout_seconds: 600
                  last_refresh:
                    started: "2026-10-08T01:10:00Z"
                    finished: "2026-10-08T01:10:02Z"
                    duration_seconds: 2.1
                    outcome: complete
                    error: null
                  last_complete_refresh: "2026-10-08T01:10:02Z"
                  next_refresh: "2026-10-08T01:15:00Z"
                  refreshes:
                    complete: 3
                    incomplete: 0
                    failed: 0
                netbox:
                  url: https://netbox.example.com/
                  up: true
                  error: null
                tracing:
                  exported: false
                  endpoint: null
                  protocol: null
        default:
          $ref: "#/components/responses/Problem"
  /server-groups:
    get:
      operationId: listServerGroups
      summary: List the server groups
      description: >-
        Returns a page of the server groups, in the configuration's order,
        each with its last-known state.
      tags: [server-groups]
      parameters:
        - $ref: "#/components/parameters/FlowID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of server groups.
          headers:
            X-Flow-ID:
              $ref: "#/components/headers/FlowID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerGroupPage"
              example:
                self: https://nbpdns.example.com/api/server-groups?limit=1
                next: https://nbpdns.example.com/api/server-groups?cursor=eyJhIjoic2l0ZS1hIn0&limit=1
                items:
                  - name: site-a
                    primary_url: https://pdns-a.example.com:8443
                    views: [_default_, internal]
                    drift_policy: report
                    status: ok
                    error: null
                    last_success: "2026-10-08T01:10:02Z"
                    counts:
                      in_sync: 980
                      drift: 15
                      missing: 5
                      inactive_in_netbox: 0
                      ignored: 1
                      unmanaged: 2
                    problem_count: 0
                    warning_count: 0
        "400":
          $ref: "#/components/responses/BadRequest"
        default:
          $ref: "#/components/responses/Problem"
  /server-groups/{group}:
    get:
      operationId: getServerGroup
      summary: Get a server group
      description: Returns one server group, with its last-known state.
      tags: [server-groups]
      parameters:
        - $ref: "#/components/parameters/FlowID"
        - $ref: "#/components/parameters/Group"
      responses:
        "200":
          description: The server group.
          headers:
            X-Flow-ID:
              $ref: "#/components/headers/FlowID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerGroup"
              example:
                name: site-a
                primary_url: https://pdns-a.example.com:8443
                views: [_default_, internal]
                drift_policy: report
                status: ok
                error: null
                last_success: "2026-10-08T01:10:02Z"
                counts:
                  in_sync: 980
                  drift: 15
                  missing: 5
                  inactive_in_netbox: 0
                  ignored: 1
                  unmanaged: 2
                problem_count: 0
                warning_count: 0
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
components:
  parameters:
    FlowID:
      name: X-Flow-ID
      in: header
      description: >-
        The request's flow ID, to follow it through the logs. A request
        without a valid one is given one. Either way, the response returns
        it, and it's the request's `request_id` in nbpdns's logs.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9._:+/=-]+$"
    Limit:
      name: limit
      in: query
      description: The most items the page may have.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 1000
        default: 100
    Cursor:
      name: cursor
      in: query
      description: >-
        Where the page starts: a previous page's `next` link carries it.
        It's opaque, and only good with the same filters as the page that
        gave it.
      schema:
        type: string
        maxLength: 1024
    Group:
      name: group
      in: path
      required: true
      description: The server group's name.
      schema:
        type: string
        pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
  headers:
    FlowID:
      description: The request's flow ID, as sent, or as nbpdns made it.
      schema:
        type: string
  responses:
    BadRequest:
      description: A parameter is invalid, such as a limit out of range, or a cursor that this API didn't give.
      headers:
        X-Flow-ID:
          $ref: "#/components/headers/FlowID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: about:blank
            title: Bad Request
            status: 400
            detail: The limit must be from 1 to 1000.
            instance: /api/server-groups
    NotFound:
      description: No such resource.
      headers:
        X-Flow-ID:
          $ref: "#/components/headers/FlowID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: about:blank
            title: Not Found
            status: 404
            detail: No server group is named site-z.
            instance: /api/server-groups/site-z
    Problem:
      description: Any other error.
      headers:
        X-Flow-ID:
          $ref: "#/components/headers/FlowID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: about:blank
            title: Internal Server Error
            status: 500
            detail: null
            instance: /api/status
  schemas:
    Problem:
      type: object
      description: An error, as RFC 9457 describes it.
      required: [type, title, status, detail, instance]
      examples:
        - type: about:blank
          title: Not Found
          status: 404
          detail: No server group is named site-z.
          instance: /api/server-groups/site-z
      properties:
        type:
          type: string
          format: uri-reference
          description: >-
            The kind of problem. `about:blank` means the status code says it
            all.
          examples: [about:blank]
        title:
          type: string
          description: A short summary of the kind of problem.
          examples: [Not Found]
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
          description: The response's HTTP status code.
          examples: [404]
        detail:
          type: [string, "null"]
          description: What went wrong with this request, or null.
          examples: [No server group is named site-z.]
        instance:
          type: [string, "null"]
          format: uri-reference
          description: The path of the request that had the problem, or null.
          examples: [/api/server-groups/site-z]
    Status:
      type: object
      description: The nbpdns service's state.
      required: [version, revision, started, uptime_seconds, ready, live, schedule, netbox, tracing]
      properties:
        version:
          type: string
          description: The build's version.
          examples: [v0.2.0]
        revision:
          type: string
          description: The build's Git commit.
          examples: [ec048d836f57f7ee1102e27b7baa38d8a17469bc]
        started:
          type: string
          format: date-time
          description: When the service started.
          examples: ["2026-10-08T01:00:00Z"]
        uptime_seconds:
          type: number
          format: double
          minimum: 0
          description: The time since the service started.
          examples: [754.2]
        ready:
          type: boolean
          description: >-
            Whether the first refresh has finished, whatever its outcome, as
            `/readyz` reports.
          examples: [true]
        live:
          type: boolean
          description: >-
            Whether a refresh has started within `drift.interval` +
            `drift.timeout` + 1 minute, as `/livez` reports.
          examples: [true]
        schedule:
          $ref: "#/components/schemas/Schedule"
        netbox:
          $ref: "#/components/schemas/NetBoxState"
        tracing:
          $ref: "#/components/schemas/Tracing"
    Schedule:
      type: object
      description: The refreshes' schedule, and how they went.
      required: [interval_seconds, timeout_seconds, last_refresh, last_complete_refresh, next_refresh, refreshes]
      properties:
        interval_seconds:
          type: number
          format: double
          minimum: 0
          description: The time from one refresh's start to the next's, `drift.interval`.
          examples: [300]
        timeout_seconds:
          type: number
          format: double
          minimum: 0
          description: How long a refresh may take, `drift.timeout`.
          examples: [600]
        last_refresh:
          description: The last refresh that finished, or null before the first.
          oneOf:
            - $ref: "#/components/schemas/Refresh"
            - type: "null"
        last_complete_refresh:
          type: [string, "null"]
          format: date-time
          description: When the last complete refresh finished, or null.
          examples: ["2026-10-08T01:10:02Z"]
        next_refresh:
          type: [string, "null"]
          format: date-time
          description: When the next refresh starts, or null while one runs.
          examples: ["2026-10-08T01:15:00Z"]
        refreshes:
          $ref: "#/components/schemas/RefreshCounts"
    Refresh:
      type: object
      description: A refresh that finished.
      required: [started, finished, duration_seconds, outcome, error]
      properties:
        started:
          type: string
          format: date-time
          description: When the refresh started.
          examples: ["2026-10-08T01:10:00Z"]
        finished:
          type: string
          format: date-time
          description: When the refresh finished.
          examples: ["2026-10-08T01:10:02Z"]
        duration_seconds:
          type: number
          format: double
          minimum: 0
          description: How long the refresh took.
          examples: [2.1]
        outcome:
          type: string
          enum: [complete, incomplete, failed]
          description: >-
            `complete` if NetBox and every group's primary were read,
            `incomplete` if a primary couldn't be, and `failed` if NetBox
            couldn't be, or the refresh ran out of time.
          examples: [complete]
        error:
          type: [string, "null"]
          description: >-
            Why a failed refresh failed: NetBox's error, or the timeout's.
            Null otherwise. A group's error is with the group.
          examples: [null]
    RefreshCounts:
      type: object
      description: The refreshes that finished since the service started, by outcome.
      required: [complete, incomplete, failed]
      properties:
        complete:
          type: integer
          format: int64
          minimum: 0
          description: Refreshes that read NetBox and every group's primary.
          examples: [3]
        incomplete:
          type: integer
          format: int64
          minimum: 0
          description: Refreshes that couldn't read a group's primary.
          examples: [0]
        failed:
          type: integer
          format: int64
          minimum: 0
          description: Refreshes that couldn't read NetBox, or ran out of time.
          examples: [0]
    NetBoxState:
      type: object
      description: NetBox's state, as of the last refresh.
      required: [url, up, error]
      properties:
        url:
          type: string
          format: uri
          description: NetBox's URL, `netbox.url`.
          examples: [https://netbox.example.com/]
        up:
          type: [boolean, "null"]
          description: Whether the last refresh could read NetBox, or null before the first.
          examples: [true]
        error:
          type: [string, "null"]
          description: Why NetBox couldn't be read, or null.
          examples: [null]
    Tracing:
      type: object
      description: Where the service's spans are exported.
      required: [exported, endpoint, protocol]
      properties:
        exported:
          type: boolean
          description: Whether spans are exported, which they are when `otlp.endpoint` is set.
          examples: [true]
        endpoint:
          type: [string, "null"]
          description: The OTLP collector's endpoint, `otlp.endpoint`, or null.
          examples: [https://otel.example.com:4318]
        protocol:
          description: The OTLP protocol, `otlp.protocol`, or null.
          oneOf:
            - $ref: "#/components/schemas/OTLPProtocol"
            - type: "null"
    OTLPProtocol:
      type: string
      description: An OTLP protocol.
      enum: [http/protobuf, grpc]
      examples: [http/protobuf]
    DriftPolicy:
      type: string
      description: >-
        What nbpdns does about a zone's drift: `report` reports it, `enforce`
        reports it and, from M13, corrects it, and `ignore` doesn't compare
        the zone.
      enum: [enforce, report, ignore]
      examples: [report]
    ServerGroup:
      type: object
      description: >-
        A PowerDNS server group, with its last-known state. The counts are
        as of `last_success`.
      required: [name, primary_url, views, drift_policy, status, error, last_success, counts, problem_count, warning_count]
      properties:
        name:
          type: string
          description: The group's name, its stable ID.
          examples: [site-a]
        primary_url:
          type: string
          format: uri
          description: The URL of the group's primary's PowerDNS API.
          examples: [https://pdns-a.example.com:8443]
        views:
          type: array
          items:
            type: string
          description: The NetBox views whose zones the group serves.
          examples: [[_default_, internal]]
        drift_policy:
          $ref: "#/components/schemas/DriftPolicy"
        status:
          type: string
          enum: [ok, failed, unknown]
          description: >-
            `ok` if the last refresh that tried the group's primary read
            it, `failed` if it couldn't, and `unknown` before any did.
          examples: [ok]
        error:
          type: [string, "null"]
          description: Why the primary couldn't be read, or null.
          examples: [null]
        last_success:
          type: [string, "null"]
          format: date-time
          description: When the group was last read and compared, or null if it never was.
          examples: ["2026-10-08T01:10:02Z"]
        counts:
          description: The group's zones by state, as of `last_success`, or null if it was never read.
          examples:
            - {in_sync: 980, drift: 15, missing: 5, inactive_in_netbox: 0, ignored: 1, unmanaged: 2}
          oneOf:
            - $ref: "#/components/schemas/ZoneCounts"
            - type: "null"
        problem_count:
          type: [integer, "null"]
          format: int64
          minimum: 0
          description: >-
            How many problems normalization worked around in either side's
            data, as of `last_success`, or null. `nbpdns drift` lists them.
          examples: [0]
        warning_count:
          type: [integer, "null"]
          format: int64
          minimum: 0
          description: >-
            How many problems with the configuration or NetBox's data the
            comparison worked around, as of `last_success`, or null.
            `nbpdns drift` lists them.
          examples: [0]
    ZoneCounts:
      type: object
      description: A server group's zones, by state.
      required: [in_sync, drift, missing, inactive_in_netbox, ignored, unmanaged]
      properties:
        in_sync:
          type: integer
          format: int64
          minimum: 0
          description: Zones that the primary serves as NetBox says.
          examples: [980]
        drift:
          type: integer
          format: int64
          minimum: 0
          description: Zones whose RRsets differ.
          examples: [15]
        missing:
          type: integer
          format: int64
          minimum: 0
          description: Active NetBox zones that the primary doesn't serve.
          examples: [5]
        inactive_in_netbox:
          type: integer
          format: int64
          minimum: 0
          description: NetBox zones that aren't active, which the primary still serves.
          examples: [0]
        ignored:
          type: integer
          format: int64
          minimum: 0
          description: Zones whose drift policy is `ignore`, which aren't compared.
          examples: [1]
        unmanaged:
          type: integer
          format: int64
          minimum: 0
          description: Zones on the primary that NetBox doesn't assign to the group. They aren't drift.
          examples: [2]
    ServerGroupPage:
      type: object
      description: One page of the server groups, with the links to it and to the next.
      required: [self, next, items]
      properties:
        self:
          type: string
          format: uri
          description: This page's absolute URL.
          examples: ["https://nbpdns.example.com/api/server-groups?limit=2"]
        next:
          type: [string, "null"]
          format: uri
          description: The next page's absolute URL, or null on the last page.
          examples: ["https://nbpdns.example.com/api/server-groups?cursor=eyJhIjoic2l0ZS1iIn0&limit=2"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/ServerGroup"
          description: The page's server groups.
          examples:
            - - name: site-a
                primary_url: https://pdns-a.example.com:8443
                views: [_default_]
                drift_policy: report
                status: unknown
                error: null
                last_success: null
                counts: null
                problem_count: null
                warning_count: null